final product

+++ app/Http/Controllers/ListingController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Listing;
+use App\Models\User;
 use Illuminate\Validation\Rule;
 
 
@@ -50,8 +51,8 @@ class ListingController extends Controller
 	    $formFields['logo'] = $request->file('logo')->store('logos', 'public');
 	}
 	//dd($formFields);
+	$formFields['user_id'] = auth()->id();
 	Listing::create($formFields);
-
 
 
 	return redirect('/')->with('message', 'Listing created successfully!');
@@ -64,6 +65,12 @@ class ListingController extends Controller
 
     //Update Listing Data
     public function update(Request $request, Listing $listing){
+	// Make sure logged in user is the owner
+
+	if($listing->user_id != auth()->id()){
+	    abort(403, 'Unauthorized Action');
+	}
+
 	$formFields = $request->validate([
 	    'title' => 'required',
 	    'company' => ['required'],
@@ -89,8 +96,15 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing){
 
+	if($listing->user_id != auth()->id()){
+	    abort(403, 'Unauthorized Action');
+	}
 	$listing->delete();
 	return redirect('/')->with('message', 'Listing Deleted Successfully.');
 
+    }
+    // Manage listings
+    public function manage(){
+	return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
